react/backend/requirements.txt
Django==5.2.8
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
psycopg2-binary==2.9.11
django-cors-headers==4.9.0
drf-spectacular==0.29.0
python-decouple==3.8
django-filter==25.2






